Background Movement
Everything now moves to the left bar the player to give the illusion the player is moving the right
rotated everything so it aligns with the x, y, z axis correctly
seperated the 'sky' and the 'mountain into two seperate assets.
Now the sky is far back, and the mountain moves left to right to make it look like its scrolling.